refactor: move StopGuard to leaf models module
bootstrap.py (service layer) imported StopGuard from wade.hooks.policies,
a service-to-hooks dependency the layering rules forbid (services may import
models/config, not hooks). StopGuard is a pure StrEnum, so relocate it to a new
leaf wade.models.hooks module and import it from there in bootstrap, hooks.cli,
and hooks.policies. Addresses CodeRabbit review on PR #378.
