+++ src/wade/hooks/policies.py
@@ -26,16 +26,15 @@ import os
 import posixpath
 import re
 import shlex
-from enum import StrEnum
 from pathlib import Path
 
 from crossby.hooks.runtime import HookDecision, HookEvent
 
+from wade.models.hooks import StopGuard
 from wade.utils import markers
 
 __all__ = [
     "GUARD_NAMES",
-    "StopGuard",
     "plan_artifact_only",
     "plan_complete",
     "session_complete",
@@ -43,21 +42,6 @@ __all__ = [
     "stop_nudge_marker_path",
     "worktree_containment",
 ]
-
-
-class StopGuard(StrEnum):
-    """The Stop-hook guards ``wade-hook stop --guard`` accepts.
-
-    A single source of truth shared by the installer
-    (:func:`wade.services.implementation_service.bootstrap._install_stop_hook`)
-    and the CLI (``wade.hooks.cli._STOP_GUARDS``), so bootstrap cannot install a
-    guard the CLI does not recognize and a typo is a type error rather than a
-    silently fail-open unknown guard. A worktree is only ever one kind of session,
-    so the two guards never share one.
-    """
-
-    SESSION_COMPLETE = "session-complete"  # impl/review sessions — nudge to run ``done``
-    PLAN_COMPLETE = "plan-complete"  # plan sessions — nudge to write a valid plan file
 
 
 # Guard names understood by the ``wade hook`` entry point. ``worktree`` / ``plan``
